src/computer-vision/requirements.txt
fastapi
pydantic
uvicorn
requests
python-multipart
pyjwt